python/agents/brand-aligner/pyproject.toml
[project]
name = "brand-aligner-agent"
version = "0.1.0"
description = "Brand Aligner Agent"
authors = [{ name = "Mohab Fekry", email = "mfekry@google.com" }]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.121.0",
  "google-adk==1.17.0",
  "google-api-python-client>=2.187.0",
  "google-auth>=2.43.0",
  "google-auth-oauthlib>=1.2.2",
  "google-cloud-aiplatform[adk,agent-engines,evaluation]<=1.126.1,>1.118.0",
  "google-cloud-storage<=3.5.0,>=2.9.1",
  "google-genai>=1.49.0",
  "matplotlib>=3.10.7",
  "numpy>=2.3.2",
  "pandas>=2.3.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.7",
    "google-adk[eval]>=1.0.0",
    "nest-asyncio>=1.6.0",
  "agent-starter-pack>=0.32.0",
    "ruff>=0.4.6",
    "mypy>=1.15.0",
    "codespell>=2.2.0",
    "types-pyyaml>=6.0.12.20240917",
    "types-requests>=2.32.0.20240914",
]

deployment = [
    "absl-py>=2.2.1",
]

  [tool.setuptools]
  packages = ["brand_aligner_agent", "deployment", "tests"]

[tool.ruff]
extend = "../../../pyproject.toml"

[tool.ruff.lint]
ignore = ["C901", "PLR0912"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["brand_aligner_agent"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
skip-magic-trailing-comma = false
docstring-code-format = true

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.agent-starter-pack]
example_question = "Evaluate this campaign for brand alignment"

[tool.agent-starter-pack.settings]
agent_directory = "brand_aligner_agent"

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true